requirements.txt
aiofiles==23.1.0
django_recaptcha==3.0.0
fastapi==0.95.0
matplotlib==3.7.1
numpy==1.24.2
orjson==3.8.9
Pillow==9.5.0
pydantic==1.10.5
SQLAlchemy==1.4.41
sqlmodel==0.0.8
uvicorn==0.21.1
